Add option to turn off swerve motor telemetry
Also adds way of turning off parts of the robot after deployment (i.e. Climber, Shooter/Intake, PhotonVision) for deployment on a bare chassis or testing of a single part. This is in response to the sluggishness of CAN ID time-outs - if a REV CAN ID is not found, the program will timeout and slow down on every loop (not good!).